Add editing and deleting features for comments

+++ multi-user-blog-kate/handlers/like.py
@@ -9,14 +9,18 @@ class Like(BlogHandler):
         post_id = self.request.get('post_id')
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         post = db.get(key)
-        liked = LikedPerson.by_person_post(self.user, post)
-        if liked:
-            post.liked_num -= 1
-            post.put()
-            db.delete(liked.key())
+        if post.author != self.user:
+            liked = LikedPerson.by_person_post(self.user, post)
+            if liked:
+                post.liked_num -= 1
+                post.put()
+                db.delete(liked.key())
+            else:
+                post.liked_num += 1
+                post.put()
+                x = LikedPerson(parent = blog_key(), person = self.user, post = post)
+                x.put()
+            self.redirect('/blog')
         else:
-            post.liked_num += 1
-            post.put()
-            x = LikedPerson(parent = blog_key(), person = self.user, post = post)
-            x.put()
-        self.redirect('/blog')
+            self.error(403)
+            return
